pyproject.toml
[tool.poetry]
name = "vending_machine"
version = "0.1.0"
description = ""
authors = ["The Gang Technology"]

[tool.poetry.dependencies]
python = ">=3.7,<4.0"
flask = "^2.2.2"
flask-sqlalchemy = "^3.0.2"


[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
flake8 = "^5.0.1"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
flake8-annotations = "^2.9.1"
flake8-eradicate = "^1.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
